Move bulb classes
