+++ resharper/src/resharper-unity/Daemon/Stages/Highlightings/UnityMarkOnGutter.cs
@@ -7,6 +7,7 @@ using JetBrains.ReSharper.Feature.Services.Bulbs;
 using JetBrains.ReSharper.Feature.Services.Daemon;
 using JetBrains.ReSharper.Feature.Services.Intentions;
 using JetBrains.ReSharper.Feature.Services.Resources;
+using JetBrains.ReSharper.Plugins.Unity.Feature.Services.Bulbs;
 using JetBrains.ReSharper.Plugins.Unity.Feature.Services.QuickFixes;
 using JetBrains.ReSharper.Psi;
 using JetBrains.ReSharper.Psi.CSharp.Tree;
