+++ resharper/src/resharper-unity/Feature/Services/Bulbs/ConvertFromCoroutineBulbAction.cs
@@ -7,13 +7,13 @@ using JetBrains.ReSharper.Psi;
 using JetBrains.ReSharper.Psi.CSharp.Tree;
 using JetBrains.TextControl;
 
-namespace JetBrains.ReSharper.Plugins.Unity.Daemon.Stages.Highlightings
+namespace JetBrains.ReSharper.Plugins.Unity.Feature.Services.Bulbs
 {
-    public class ConvertToCoroutineBulbAction : BulbActionBase
+    public class ConvertFromCoroutineBulbAction : BulbActionBase
     {
         private readonly IMethodDeclaration myMethodDeclaration;
 
-        public ConvertToCoroutineBulbAction(IMethodDeclaration methodDeclaration)
+        public ConvertFromCoroutineBulbAction(IMethodDeclaration methodDeclaration)
         {
             myMethodDeclaration = methodDeclaration;
         }
@@ -23,16 +23,20 @@ namespace JetBrains.ReSharper.Plugins.Unity.Daemon.Stages.Highlightings
             var element = myMethodDeclaration.DeclaredElement;
             if (element == null) return null;
 
-            var predefinedTypeCache = solution.GetComponent<IPredefinedTypeCache>();
-            var predefinedType = predefinedTypeCache.GetOrCreatePredefinedType(myMethodDeclaration.GetPsiModule());
+            var unityApi = solution.GetComponent<UnityApi>();
+            var eventFunction = unityApi.GetUnityEventFunction(element);
+
+            IType returnType = TypeFactory.CreateTypeByCLRName(eventFunction.ReturnType, myMethodDeclaration.GetPsiModule());
+            if (eventFunction.ReturnTypeIsArray)
+                returnType = TypeFactory.CreateArrayType(returnType, 1);
 
             var language = myMethodDeclaration.Language;
             var changeTypeHelper = LanguageManager.Instance.GetService<IChangeTypeHelper>(language);
-            changeTypeHelper.ChangeType(predefinedType.IEnumerator, element);
+            changeTypeHelper.ChangeType(returnType, element);
 
             return null;
         }
 
-        public override string Text => "To coroutine";
+        public override string Text => "To standard event function";
     }
 }
