+++ resharper/src/resharper-unity/Feature/Services/Bulbs/ConvertToCoroutineBulbAction.cs
@@ -7,13 +7,13 @@ using JetBrains.ReSharper.Psi;
 using JetBrains.ReSharper.Psi.CSharp.Tree;
 using JetBrains.TextControl;
 
-namespace JetBrains.ReSharper.Plugins.Unity.Daemon.Stages.Highlightings
+namespace JetBrains.ReSharper.Plugins.Unity.Feature.Services.Bulbs
 {
-    public class ConvertFromCoroutineBulbAction : BulbActionBase
+    public class ConvertToCoroutineBulbAction : BulbActionBase
     {
         private readonly IMethodDeclaration myMethodDeclaration;
 
-        public ConvertFromCoroutineBulbAction(IMethodDeclaration methodDeclaration)
+        public ConvertToCoroutineBulbAction(IMethodDeclaration methodDeclaration)
         {
             myMethodDeclaration = methodDeclaration;
         }
@@ -23,20 +23,16 @@ namespace JetBrains.ReSharper.Plugins.Unity.Daemon.Stages.Highlightings
             var element = myMethodDeclaration.DeclaredElement;
             if (element == null) return null;
 
-            var unityApi = solution.GetComponent<UnityApi>();
-            var eventFunction = unityApi.GetUnityEventFunction(element);
-
-            IType returnType = TypeFactory.CreateTypeByCLRName(eventFunction.ReturnType, myMethodDeclaration.GetPsiModule());
-            if (eventFunction.ReturnTypeIsArray)
-                returnType = TypeFactory.CreateArrayType(returnType, 1);
+            var predefinedTypeCache = solution.GetComponent<IPredefinedTypeCache>();
+            var predefinedType = predefinedTypeCache.GetOrCreatePredefinedType(myMethodDeclaration.GetPsiModule());
 
             var language = myMethodDeclaration.Language;
             var changeTypeHelper = LanguageManager.Instance.GetService<IChangeTypeHelper>(language);
-            changeTypeHelper.ChangeType(returnType, element);
+            changeTypeHelper.ChangeType(predefinedType.IEnumerator, element);
 
             return null;
         }
 
-        public override string Text => "To standard event function";
+        public override string Text => "To coroutine";
     }
 }
